requirements.txt
streamlit>=1.28.1
PyPDF2==3.0.1
# Use a faiss-cpu wheel available on PyPI for Windows/Python 3.13
faiss-cpu>=1.9.0
google-generativeai>=0.8.0
groq>=0.4.2
python-dotenv>=1.0.0
numpy>=1.24.3
